Set termios settings immediately when specified.
In a test application that opens two ttyMAX ports and sends data between them:
cs tx     rx
1: cc01 - 0000 # set write conf reg
1: 80aa - 4200 # send aa
1: 0000 - 4200 # No more data to read
<switch to cs2>
2: cc01 - 0000 # set conf register
2: 0000 - 4200 # read returns nothing

It opens both UART /dev/ descriptors, sets up termios, tcdrains, and then attempts to send on cs 1, and receive on cs2.  The current driver waits to configure the port until port_dowork is called, so without this patch when it sets the conf register it clears the fifo and the data we were waiting to read is gone.