ARM: ts4900_yocto: Added Yocto Jethro defconfig
Includes FT5x06 driver as a builtin

Changes CONFIG_MXC_GPU_VIV to a module. When installed over Jethro
image, it is required that this kernel's galcore.ko be moved out
of the way, and the Yocto built galcore.ko copied in its place.
Youcto Jethro builds this module out of kernel and all applications
expect to have that driver version available. Since its the same
kernel rev, the Yocto built module can be loaded in this kernel
without issue. Doing so will allow X to start properly.

Retrived from:
https://raw.githubusercontent.com/embeddedarm/meta-ts/jethro/recipes-kernel/linux/linux-technologic-4.1.15/defconfig